ran the model maker again

+++ server/app/models/budget.rb
@@ -4,12 +4,12 @@ class Budget
 
   embedded_in :phase
 
-  field :initial_phase_value_total, type: Double
-  field :final_phase_value_total, type: Double
-  field :initial_phase_fee_total, type: Double
-  field :final_phase_fee_total, type: Double
-  field :construction_initial, type: Double
-  field :construction_final, type: Double
+  field :initial_phase_value_total, type: Float
+  field :final_phase_value_total, type: Float
+  field :initial_phase_fee_total, type: Float
+  field :final_phase_fee_total, type: Float
+  field :construction_initial, type: Float
+  field :construction_final, type: Float
   field :comments, type: String
   field :fee_type, type: String
 end
