+++ server/app/models/geo_json.rb
@@ -4,7 +4,7 @@ class GeoJson
 
   embedded_in :location
 
-  field :lat, type: Double
-  field :lon, type: Double
+  field :lat, type: Float
+  field :lon, type: Float
   field :boundary_points, type: Array
 end
