+++ server/app/models/gross_building_area.rb
@@ -4,7 +4,7 @@ class GrossBuildingArea
 
   embedded_in :dimension_group, inverse_of: :gba
 
-  field :measurement, type: Double
+  field :measurement, type: Float
   field :standard_name, type: String
   field :standard_version, type: String
 end
