+++ server/app/models/net_lettable_area.rb
@@ -4,7 +4,7 @@ class NetLettableArea
 
   embedded_in :dimension_group, inverse_of: :nla
 
-  field :measurement, type: Double
+  field :measurement, type: Float
   field :standard_name, type: String
   field :standard_version, type: String
 end
